+++ src/components/Avatar/Avatar.stories.tsx
@@ -13,7 +13,7 @@ export default {
   parameters: {
     layout: 'centered',
   },
-  tags: ['autodocs', 'version:2.0.1'],
+  tags: ['autodocs', 'version:2.1'],
 } as Meta<typeof Avatar>;
 
 type Story = StoryObj<typeof Avatar>;
@@ -24,6 +24,9 @@ export const Small: Story = {
   args: {
     size: 'sm',
     variant: 'icon',
+    user: {
+      fullName: 'Douglas Adams',
+    },
   },
 };
 
@@ -31,6 +34,9 @@ export const SmallText: Story = {
   args: {
     ...Small.args,
     variant: 'text',
+    user: {
+      fullName: 'Stephen Baxter',
+    },
   },
 };
 
@@ -38,6 +44,9 @@ export const Medium: Story = {
   args: {
     size: 'md',
     variant: 'icon',
+    user: {
+      fullName: 'Ray Bradbury',
+    },
   },
 };
 
@@ -45,6 +54,9 @@ export const MediumText: Story = {
   args: {
     ...Medium.args,
     variant: 'text',
+    user: {
+      fullName: 'Robert Heinlein',
+    },
   },
 };
 
@@ -52,6 +64,9 @@ export const Large: Story = {
   args: {
     size: 'lg',
     variant: 'icon',
+    user: {
+      fullName: 'Arthur C. Clarke',
+    },
   },
 };
 
@@ -59,6 +74,9 @@ export const LargeText: Story = {
   args: {
     ...Large.args,
     variant: 'text',
+    user: {
+      fullName: 'Michael Crichton',
+    },
   },
 };
 
@@ -66,6 +84,9 @@ export const ExtraLarge: Story = {
   args: {
     size: 'xl',
     variant: 'icon',
+    user: {
+      fullName: 'Frank Herbert',
+    },
   },
 };
 
@@ -73,6 +94,9 @@ export const ExtraLargeText: Story = {
   args: {
     ...ExtraLarge.args,
     variant: 'text',
+    user: {
+      fullName: 'Edward M. Forster',
+    },
   },
 };
 
@@ -185,4 +209,16 @@ export const UsingWithTooltip: Story = {
       </span>
     </Tooltip>
   ),
+};
+
+/**
+ * When using color = `fixed` `Avatar` will use a single, static color, which doesn't change based on the user data supplied.
+ *
+ * Colors can be changed by overriding the `--avatar__fg`, `--avatar__bg`, and `--avatar__border` CSS Properties.
+ */
+export const Fixed: Story = {
+  args: {
+    ...Default.args,
+    color: 'fixed',
+  },
 };